replace entry folder with index, renamed bundle file, updated sass logic and exports
