+++ src/_config/styles/theme/_theme.module.scss
@@ -2,11 +2,12 @@
 // Note: this file used to add sass variables into components
 @use './_theme.scss' as theme;
 
-// Use sass:map to get values from $exported map
+// Use sass:map and sass:meta to export values
 @use "sass:map";
+@use "sass:meta";
 
 // Map module-variables from theme file
-$exported: module-variables(theme);
+$exported: meta.module-variables(theme);
 
 :export {
 	bp01 : map.get($exported, 'bp-01');
