+++ src/components/slideout/Slideout.jsx
@@ -8,7 +8,7 @@ import './styles/slideout.scss';
 import { slideout } from './scripts/slideout';
 
 /* Local components */
-import { Context } from '../../entry/context/Context';
+import { Context } from '../../index/Context';
 
 export const Slideout = (props) => {
 	let { id, width, direction, label, content } = props;
