+++ src/index.html
@@ -537,6 +537,6 @@
 		</svg>
 		<div id="root"></div>
 		<div id="portal"></div>
-		<script type="module" src="./entry/index/Index.jsx"></script>
+		<script type="module" src="./index/Index.jsx"></script>
 	</body>
 </html>
