+++ src/index/Context.jsx
@@ -1,4 +1,5 @@
 /* React */
 import { createContext } from 'react';
 
+/* Create Context.Provider wrapper */
 export const Context = createContext();
